requirements.txt
torch==1.10.1          # PyTorch for deep learning
torchvision==0.11.2    # TorchVision for image transformations and pretrained models
numpy==1.21.4          # Numerical computations
opencv-python==4.5.4.60 # OpenCV for video and image processing
matplotlib==3.5.0      # Visualization library
scipy==1.7.3           # Scientific computing (optional)
scikit-learn==1.0.1    # For metrics or clustering (if needed)
Pillow==8.4.0          # Image processing
tqdm==4.62.3           # Progress bars
numba==0.54.1          # For performance optimizations (optional)
pyyaml==6.0            # YAML configuration files (optional)
pandas==1.3.4          # Data manipulation
deepsort==1.4.0        # DeepSORT for tracking (customizable)
